Updated BFQ to v2. (cherry-picking from commit ba9287f45bed3d9dca621c6b0524660002e5efe3) Switch from BFQ-v2 to BFQ-v2-r1.
Improvements:

When the first new request for the active queue arrives during disk idling,
the device is not immediately unplugged if this request is small. This prevents
the disk to be immediately committed to serve a small request. In fact, further
requests are likely to be merged to this one quickly, after which the device is
unplugged by some block layer function and larger requests are dispatched.

The idling timeout and the weight-raising max time have been tuned a little bit
too, as detailed below.

Reduced a little bit the disk idling timeout for weight-raised
queues. This helps reduce latencies for applications made of several
processes that do something and then pass the token to the next
process. By reducing the timeout we reduce the time that may be wasted
to wait for the next request of a process that has just passed the token,
and hence that will not issue any new request.

Reduced the weight-raising max time from 8 to 7.5 seconds: this
reduces a little bit the time interval during which a just-started
disk-bound process may unproperly be weight-raised. (cherry-picking from commit ee37cba527c3a81eae1551123c507cbc63e6a200)
Updated BFQ to v2. (cherry-picking from commit ba9287f45bed3d9dca621c6b0524660002e5efe3)